Backend/requirements.txt
fastapi==0.114.0
uvicorn==0.24.0
motor==3.5.1
pydantic==2.9.0
python-dotenv==1.0.0